run correct number of cycles per second

+++ core/src/cycles.rs
@@ -3,12 +3,12 @@ use derive_more::{Add, AddAssign, Display, From, Sub, SubAssign};
 #[derive(
     Debug, Display, Clone, Copy, PartialEq, PartialOrd, Add, AddAssign, Sub, SubAssign, From,
 )]
-pub struct MCycles(#[display(fmt = "{} m-cycles", _0)] pub u16);
+pub struct MCycles(#[display(fmt = "{} m-cycles", _0)] pub u32);
 
 #[derive(
     Debug, Display, Clone, Copy, PartialEq, PartialOrd, Add, AddAssign, Sub, SubAssign, From,
 )]
-pub struct TCycles(#[display(fmt = "{} t-cycles", _0)] pub u16);
+pub struct TCycles(#[display(fmt = "{} t-cycles", _0)] pub u32);
 
 impl From<MCycles> for TCycles {
     fn from(m: MCycles) -> Self {
